% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{new_sdtm_block}
\alias{new_sdtm_block}
\title{Adam block}
\usage{
new_sdtm_block(selected, ...)
}
\arguments{
\item{selected}{Default selected dataset. Must belong to the
pharmaversesdtm package.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}}
}
\description{
Block listing pharmaversesdtm
}
